feat: add search query
