+++ src/features/Search/ui/Search.tsx
@@ -4,18 +4,18 @@ import { Button, LOCAL_SEARCH, useGetLocalData } from '@/shared';
 import { Input } from '@/shared/ui/Input/Input';
 import { FC, KeyboardEvent, useEffect, useState } from 'react';
 import cls from './Search.module.css';
+import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 
 interface SearchProps {
   className?: string;
 }
 
-interface LocalStorageChangedEvent extends Event {
-  newValue: string;
-}
-
 export const Search: FC<SearchProps> = (props) => {
   const [error, setError] = useState('');
   const { setValue, value } = useGetLocalData();
+  const params = useSearchParams();
+  const router = useRouter();
+  const pathname = usePathname();
 
   const classes = [cls.block, props.className].join(' ');
 
@@ -28,12 +28,14 @@ export const Search: FC<SearchProps> = (props) => {
 
   const saveLocal = () => {
     if (value !== undefined) {
-      const event = new Event(
-        'localStorageChanged'
-      ) as LocalStorageChangedEvent;
-      event.newValue = value.trim();
-      localStorage.setItem(LOCAL_SEARCH, event.newValue);
-      window.dispatchEvent(event);
+      const query = value.trim();
+
+      localStorage.setItem(LOCAL_SEARCH, query);
+      const searchParams = new URLSearchParams(params);
+      searchParams.set('query', query);
+      searchParams.set('page', '1');
+
+      router.push(`${pathname}?${searchParams.toString()}`);
     }
   };
